提交代码

+++ Modules/module_widget/src/main/java/com/example/william/my/module/widget/adapter/RecyclerAdapter.java
@@ -42,7 +42,7 @@ public class RecyclerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolde
      */
     @Override
     public long getItemId(int position) {
-        return (long) position;
+        return position;
     }
 
     private static class ViewHolder extends RecyclerView.ViewHolder {
